requirements.txt
# YAML
PyYAML
tqdm